pyproject.toml
[project]
name = "ai-search-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.19.0",
    "azure-search-documents==11.6.0b1",
    "python-dotenv>=1.0.1",
]
